Скрипт для меню сортировки, активный фильтр, модули по IIFE и ручному подключению
