+++ src/catalog.html
@@ -227,7 +227,15 @@
       </div>
     </li>
   </template>
-  <script src="js/main.min.js"></script>
+  <script src="js/backend.js"></script>
+  <script src="js/product.js"></script>
+  <script src="js/catalog.js"></script>
+  <script src="js/data.js"></script>
+  <script src="js/modalpopup.js"></script>
+  <script src="js/filterform.js"></script>
+  <script src="js/sortmenu.js"></script>
+  <script src="js/rangeControls.js"></script>
+  <script src="js/main.js"></script>
 </body>
 
 </html>
