+++ src/index.html
@@ -278,7 +278,9 @@
     </form>
     <button class="modal__close" type="button" name="closemodal"><span class="visually-hidden">Закрыть</span></button>
   </section>
-  <script src="js/main.min.js"></script>
+  <script src="js/modalpopup.js"></script>
+  <script src="js/slider.js"></script>
+  <script src="js/main.js"></script>
 </body>
 
 </html>
