+++ src/js/backend.js
@@ -1,0 +1,58 @@
+(function () {
+
+  const URL_LOAD = 'js/products.json';
+  const URL_UPLOAD = '';
+  const STATUS_CODE_OK = 200;
+  const TIMEOUT_IN_MS = 10000;
+
+  function load(onSuccess, onError) {
+    const xhr = new XMLHttpRequest();
+    xhr.responseType = 'json';
+    xhr.timeout = TIMEOUT_IN_MS;
+
+    xhr.addEventListener('load', function () {
+      if (xhr.status === STATUS_CODE_OK) {
+        onSuccess(xhr.response);
+      } else {
+        onError('Ошибка при загрузке данных. Статус ответа: ' + xhr.status + ' ' + xhr.statusText);
+      }
+    });
+    xhr.addEventListener('error', function () {
+      onError('Произошла ошибка при загрузке данных. ');
+    });
+    xhr.addEventListener('timeout', function () {
+      onError('Ошибка при загрузке данных. Запрос не успел выполниться за ' + xhr.timeout + 'мс');
+    });
+
+    xhr.open('GET', URL_LOAD);
+    xhr.send();
+  }
+
+  function upload(data, onLoad, onError) {
+    const xhr = new XMLHttpRequest();
+    xhr.responseType = 'json';
+    xhr.timeout = TIMEOUT_IN_MS;
+
+    xhr.addEventListener('load', function () {
+      if (xhr.status === STATUS_CODE_OK) {
+        onLoad(xhr.response);
+      } else {
+        onError();
+      }
+    });
+    xhr.addEventListener('error', function () {
+      onError();
+    });
+    xhr.addEventListener('timeout', function () {
+      onError();
+    });
+
+    xhr.open('POST', URL_UPLOAD);
+    xhr.send(data);
+  }
+
+   window.backend = {
+    load: load,
+    upload: upload
+  };
+})();
